requirements.txt
amqp==5.0.9
asgiref==3.5.0
billiard==3.6.4.0
black==22.1.0
celery==5.2.3
click==8.0.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
colorama==0.4.4
Deprecated==1.2.13
Django==4.0.2
django-filter==21.1
djangorestframework==3.13.1
flake8==4.0.1
importlib-metadata==4.10.1
kombu==5.2.3
Markdown==3.3.6
mccabe==0.6.1
mypy-extensions==0.4.3
packaging==21.3
pathspec==0.9.0
platformdirs==2.4.1
prompt-toolkit==3.0.26
psycopg2==2.9.3
pycodestyle==2.8.0
pyflakes==2.4.0
pyparsing==3.0.7
pytz==2021.3
redis==4.1.2
six==1.16.0
sqlparse==0.4.2
tomli==2.0.0
typing_extensions==4.0.1
tzdata==2021.5
vine==5.0.0
wcwidth==0.2.5
wrapt==1.13.3
zipp==3.7.0
